requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.9
astunparse==1.6.3
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
autopep8==2.3.2
beautifulsoup4==4.14.2
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
crashtest==0.4.1
cryptography==44.0.2
cssselect==1.3.0
daphne==4.1.2
decorator==5.2.1
deepface==0.0.95
dicttoxml==1.7.16
dill==0.4.0
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
docutils==0.21.2
drf-spectacular==0.28.0
dulwich==0.22.8
et_xmlfile==2.0.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
frozenlist==1.6.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
geocoder==1.38.1
geoip2==5.0.1
google-pasta==0.2.0
grpcio==1.75.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.8
httpx==0.28.1
hyperlink==21.0.0
id==1.5.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
installer==0.7.0
ip2geotools==0.1.6
IP2Location==8.10.5
isort==6.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.11.3
keyring==25.6.0
kombu==5.5.3
lazy-object-proxy==1.11.0
libclang==18.1.1
lxml==5.3.2
lz4==4.4.4
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
maxminddb==2.6.3
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.3
more-itertools==10.6.0
msgpack==1.1.0
mtcnn==1.0.0
multidict==6.4.3
namex==0.1.0
nh3==0.2.21
num2words==0.5.14
numpy==2.2.5
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pbs-installer==2025.4.9
pillow==11.2.1
pip-review==1.3.0
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
poetry-plugin-export==1.9.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.32.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pylint==3.3.6
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
ratelim==0.1.6
readme_renderer==44.0
redis==5.2.1
referencing==0.36.2
reportlab==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3986==2.0.0
rich==14.0.0
rpds-py==0.24.0
SecretStorage==3.3.3
selenium==4.31.0
service-identity==24.2.0
setuptools==79.0.0
shellingham==1.5.4
six==1.17.0
slugify==0.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tomlkit==0.13.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
trove-classifiers==2025.4.11.15
twine==6.1.0
Twisted==24.11.0
txaio==23.1.1
typed_ast==1.5.5
typing==3.7.4.3
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
unicode==2.9
uritemplate==4.1.1
urllib3==2.4.0
user-agents==2.2.0
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zope.interface==7.2
zstandard==0.23.0
